requirements.txt
lz4==3.1.3
matplotlib==3.3.4
msgpack==1.0.2
numpy==1.19.5
pandas==1.1.5
psutil==5.8.0
psycopg2-binary==2.9.1
pynacl==1.4.0
pycryptodome==3.10.1
pympler==0.9
pyspark==3.1.2
redis==3.5.3
python-snappy==0.6.0
sqlalchemy==1.4.22
sqlparse==0.4.1
treelib==1.6.1
zstd==1.5.0.2
